feat: attempt to simplify & improve reliability of user events
We have event handlers that trigger additional tasks when a user account or user
group are changed. Primarily this is to trigger some adjustment to the users
corresponding Microsoft 365 account or Sharepoint access.

This change attempts to make creating the 365 account and assigning a users 365
licence more reliable when various user events occur - particularly when the
account is first created.

Assigning a user's initial 365 licence is made more reliable by retrying on
failure to allow for eventual consistency in Microsoft Graph after account
creation. This resolves TEC-1871.

We also trigger attempting to create a 365 account if it doesn't already exist
in some additional circumstances like when a user is activated or their role
changes.

We have also added some additional complications like storing the users initial
password (encrypted) in their user account so we can access it later to send the
user a welcome email including their 365 password. We want to send the welcome
message asynchronously but don't want the password sent to background task
manager and potentially ended up in logs.

NB A lot of this stuff can be removed or simplified if we establish
auto-provisioning of user accounts.

+++ app/accounts/access.py
@@ -13,16 +13,19 @@ from typing import TYPE_CHECKING, Protocol
 
 from core.models import Issue
 
-from accounts.models import CaseGroups
 
 if TYPE_CHECKING:
-    from django.contrib.auth.models import Group
-
     from accounts.models import User
 
 
 class UserAccessEventManagerInterface(Protocol):
     """Abstract interface for user access operations."""
+
+    def user_activated(self, user: User) -> None:
+        raise NotImplementedError()
+
+    def user_deactivated(self, user: User) -> None:
+        raise NotImplementedError()
 
     def user_added_to_case(self, user: User, issue: Issue) -> None:
         raise NotImplementedError()
@@ -30,16 +33,7 @@ class UserAccessEventManagerInterface(Protocol):
     def user_removed_from_case(self, user: User, issue: Issue) -> None:
         raise NotImplementedError()
 
-    def user_added_to_group(self, user: User, group: Group) -> None:
-        raise NotImplementedError()
-
-    def user_removed_from_group(self, user: User, group: Group) -> None:
-        raise NotImplementedError()
-
-    def user_activated(self, user: User) -> None:
-        raise NotImplementedError()
-
-    def user_deactivated(self, user: User) -> None:
+    def user_role_changed(self, user: User) -> None:
         raise NotImplementedError()
 
 
@@ -53,30 +47,32 @@ class UserAccessEventAdapter(UserAccessEventManagerInterface):
 
         self._service = ms_service
 
+    def user_activated(self, user: User) -> None:
+        if user.role.is_paralegal_or_better:
+            self._service.set_up_new_user(user)
+            self._service.add_office_licence(user)
+
+        if user.role.is_coordinator_or_better:
+            self._service.add_group_member(user)
+
+    def user_deactivated(self, user: User) -> None:
+        self._service.remove_group_member(user)
+        self._service.remove_office_licence(user)
+
     def user_added_to_case(self, user: User, issue: Issue) -> None:
         self._service.add_user_to_case(user, issue)
 
     def user_removed_from_case(self, user: User, issue: Issue) -> None:
         self._service.remove_user_from_case(user, issue)
 
-    def user_added_to_group(self, user: User, group: Group) -> None:
-        if group.name == CaseGroups.COORDINATOR or group.name == CaseGroups.ADMIN:
-            self._service.add_group_member(user)
+    def user_role_changed(self, user: User) -> None:
+        if not user.role.is_paralegal_or_better:
+            self._service.remove_office_licence(user)
+        else:
+            self._service.set_up_new_user(user)
+            self._service.add_office_licence(user)
 
-    def user_removed_from_group(self, user: User, group: Group) -> None:
-        # Remove the users group membership when they no longer have coordinator
-        # or admin roles.
-        if (
-            group.name == CaseGroups.COORDINATOR or group.name == CaseGroups.ADMIN
-        ) and not user.role.is_coordinator_or_better:
+        if not user.role.is_coordinator_or_better:
             self._service.remove_group_member(user)
-
-    def user_activated(self, user: User) -> None:
-        self._service.add_office_licence(user)
-        if user.role.is_coordinator_or_better:
+        else:
             self._service.add_group_member(user)
-
-    def user_deactivated(self, user: User) -> None:
-        if user.role.is_coordinator_or_better:
-            self._service.remove_group_member(user)
-        self._service.remove_office_licence(user)
